Refactors code into more manageable parts
