+++ flask_restless/views/base.py
@@ -57,37 +57,29 @@ from mimerender import FlaskMimeRender
 from mimerender import register_mime
 from sqlalchemy.exc import DataError
 from sqlalchemy.exc import IntegrityError
-from sqlalchemy.exc import OperationalError
 from sqlalchemy.exc import ProgrammingError
 from sqlalchemy.orm.exc import FlushError
 from sqlalchemy.orm.exc import MultipleResultsFound
 from sqlalchemy.orm.exc import NoResultFound
 from werkzeug import parse_options_header
-from werkzeug.exceptions import BadRequest
 from werkzeug.exceptions import HTTPException
 
-from .helpers import changes_on_update
-from .helpers import collection_name
+from ..helpers import collection_name
+from ..helpers import get_model
+from ..helpers import is_like_list
+from ..helpers import primary_key_name
+from ..helpers import primary_key_value
+from ..helpers import url_for
+from ..search import ComparisonToNull
+from ..search import search
+from ..search import search_relationship
+from ..serialization import simple_serialize
+from ..serialization import simple_relationship_serialize
+from ..serialization import DefaultDeserializer
+from ..serialization import DeserializationException
+from ..serialization import SerializationException
 from .helpers import count
-from .helpers import evaluate_functions
-from .helpers import get_by
-from .helpers import get_model
-from .helpers import get_related_model
-from .helpers import has_field
-from .helpers import is_like_list
-from .helpers import primary_key_name
-from .helpers import primary_key_value
-from .helpers import strings_to_datetimes
 from .helpers import upper_keys as upper
-from .helpers import url_for
-from .search import ComparisonToNull
-from .search import search
-from .search import search_relationship
-from .serialization import simple_serialize
-from .serialization import simple_relationship_serialize
-from .serialization import DefaultDeserializer
-from .serialization import DeserializationException
-from .serialization import SerializationException
 
 #: String used internally as a dictionary key for passing header information
 #: from view functions to the :func:`jsonpify` function.
@@ -905,45 +897,6 @@ class ModelView(MethodView):
         self.model = model
 
 
-class FunctionAPI(ModelView):
-    """Provides method-based dispatching for :http:method:`get` requests which
-    wish to apply SQL functions to all instances of a model.
-
-    .. versionadded:: 0.4
-
-    """
-
-    def get(self):
-        """Returns the result of evaluating the SQL functions specified in the
-        body of the request.
-
-        For a description of the request and response formats, see
-        :ref:`functionevaluation`.
-
-        """
-        if 'functions' not in request.args:
-            detail = 'Must provide `functions` query parameter'
-            return error_response(400, detail=detail)
-        functions = request.args.get('functions')
-        try:
-            data = json.loads(str(functions)) or []
-        except (TypeError, ValueError, OverflowError) as exception:
-            detail = 'Unable to decode JSON in `functions` query parameter'
-            return error_response(400, cause=exception, detail=detail)
-        try:
-            result = evaluate_functions(self.session, self.model, data)
-        except AttributeError as exception:
-            detail = 'No such field "{0}"'.format(exception.field)
-            return error_response(400, cause=exception, detail=detail)
-        except KeyError as exception:
-            detail = str(exception)
-            return error_response(400, cause=exception, detail=detail)
-        except OperationalError as exception:
-            detail = 'No such function "{0}"'.format(exception.function)
-            return error_response(400, cause=exception, detail=detail)
-        return dict(data=result)
-
-
 class APIBase(ModelView):
     """Base class for view classes that provide fetch, create, update, and
     delete functionality for resources and relationships on resources.
@@ -1461,946 +1414,3 @@ class APIBase(ModelView):
             toinclude = set(toinclude.split(','))
         return set(chain(resources_from_path(instance, path)
                          for path in toinclude))
-
-
-class API(APIBase):
-    """Provides method-based dispatching for :http:method:`get`,
-    :http:method:`post`, :http:method:`patch`, and :http:method:`delete`
-    requests, for both collections of resources and individual resources.
-
-    `session` and `model` are as described in the constructor of the
-    superclass. In addition to those described below, this constructor also
-    accepts all the keyword arguments of the constructor of the superclass.
-
-    `page_size`, `max_page_size`, `serializer`, `deserializer`, `includes`, and
-    `allow_client_generated_ids` are as described in
-    :meth:`APIManager.create_api`.
-
-    """
-
-    def __init__(self, session, model, allow_client_generated_ids=False, *args,
-                 **kw):
-        super(API, self).__init__(session, model, *args, **kw)
-
-        #: Whether this API allows the client to specify the ID for the
-        #: resource to create; for more information, see
-        #: :ref:`clientids`.
-        self.allow_client_generated_ids = allow_client_generated_ids
-
-        #: Whether any side-effect changes are made to the SQLAlchemy
-        #: model on updates.
-        self.changes_on_update = changes_on_update(self.model)
-
-    def collection_processor_type(self, is_relation=False, **kw):
-        """The suffix for the pre- and postprocessor identifiers for
-        requests on collections of resources.
-
-        `is_relation` is ``True`` if and only if the request is for a
-        to-many relation. Otherwise, the request is for a collection of
-        primary resources.
-
-        """
-        return 'TO_MANY_RELATION' if is_relation else 'COLLECTION'
-
-    def resource_processor_type(self, is_relation=False, **kw):
-        """The suffix for the pre- and postprocessor identifiers for
-        requests on a single resource.
-
-        `is_relation` is ``True`` if and only if the request is for a
-        to-one relation. Otherwise, the request is for a single
-        resource.
-
-        """
-        return 'TO_ONE_RELATION' if is_relation else 'RESOURCE'
-
-    def _get_related_resource(self, resource_id, relation_name,
-                              related_resource_id):
-        """Returns a response containing a resource related to a given
-        resource.
-
-        For example, a request like this::
-
-            GET /people/1/articles/2
-
-        will fetch the article with ID 2 that is related to the person with ID
-        1 via the ``articles`` relationship. In general, this method is called
-        on requests of the form::
-
-            GET /<collection_name>/<resource_id>/<relation_name>/<related_resource_id>
-
-        """
-        for preprocessor in self.preprocessors['GET_RELATED_RESOURCE']:
-            temp_result = preprocessor(resource_id=resource_id,
-                                       relation_name=relation_name,
-                                       related_resource_id=related_resource_id)
-            # Let the return value of the preprocessor be the new value of
-            # instid, thereby allowing the preprocessor to effectively specify
-            # which instance of the model to process on.
-            #
-            # We assume that if the preprocessor returns None, it really just
-            # didn't return anything, which means we shouldn't overwrite the
-            # instid.
-            if temp_result is not None:
-                if isinstance(temp_result, tuple):
-                    if len(temp_result) == 2:
-                        resource_id, relation_name = temp_result
-                    else:
-                        resource_id, relation_name, related_resource_id = \
-                            temp_result
-                else:
-                    resource_id = temp_result
-        # Get the resource with the specified ID.
-        primary_resource = get_by(self.session, self.model, resource_id,
-                                  self.primary_key)
-        # Return an error if there is no resource with the specified ID.
-        if primary_resource is None:
-            detail = 'No instance with ID {0}'.format(resource_id)
-            return error_response(404, detail=detail)
-        # Return an error if the relation is a to-one relation.
-        if not is_like_list(primary_resource, relation_name):
-            detail = ('Cannot access a related resource by ID from a to-one'
-                      ' relation')
-            return error_response(404, detail=detail)
-        # Get the model of the specified relation.
-        related_model = get_related_model(self.model, relation_name)
-        # Return an error if no such relation exists.
-        if related_model is None:
-            detail = 'No such relation: {0}'.format(related_model)
-            return error_response(404, detail=detail)
-        # Get the related resources.
-        resources = getattr(primary_resource, relation_name)
-        # Check if one of the related resources has the specified ID. (JSON API
-        # expects all IDs to be strings.)
-        primary_keys = (primary_key_value(i) for i in resources)
-        if not any(str(k) == related_resource_id for k in primary_keys):
-            detail = 'No related resource with ID {0}'
-            detail = detail.format(related_resource_id)
-            return error_response(404, detail=detail)
-        # Get the related resource by its ID.
-        resource = get_by(self.session, related_model, related_resource_id)
-        return self._get_resource_helper(resource,
-                                         primary_resource=primary_resource,
-                                         relation_name=relation_name,
-                                         related_resource=True)
-
-    def _get_relation(self, resource_id, relation_name):
-        """Returns a response containing a resource or a collection of
-        resources related to a given resource.
-
-        For example, a request for a to-many relationship like this::
-
-            GET /people/1/articles
-
-        will fetch the articles related to the person with ID 1 via the
-        ``articles`` relationship. On a request to a to-one relationship::
-
-            GET /articles/2/author
-
-        a single resource will be returned.
-
-        In general, this method is called on requests of the form::
-
-            GET /<collection_name>/<resource_id>/<relation_name>
-
-        """
-        try:
-            filters, sort, group_by, single = self._collection_parameters()
-        except (TypeError, ValueError, OverflowError) as exception:
-            detail = 'Unable to decode filter objects as JSON list'
-            return error_response(400, cause=exception, detail=detail)
-        except SingleKeyError as exception:
-            detail = 'Invalid format for filter[single] query parameter'
-            return error_response(400, cause=exception, detail=detail)
-
-        for preprocessor in self.preprocessors['GET_RELATION']:
-            temp_result = preprocessor(resource_id=resource_id,
-                                       relation_name=relation_name,
-                                       filters=filters, sort=sort,
-                                       group_by=group_by, single=single)
-            # Let the return value of the preprocessor be the new value of
-            # instid, thereby allowing the preprocessor to effectively specify
-            # which instance of the model to process on.
-            #
-            # We assume that if the preprocessor returns None, it really just
-            # didn't return anything, which means we shouldn't overwrite the
-            # instid.
-            if temp_result is not None:
-                if isinstance(temp_result, tuple) and len(temp_result) == 2:
-                    resource_id, relation_name = temp_result
-                else:
-                    resource_id = temp_result
-
-        # Get the resource with the specified ID.
-        primary_resource = get_by(self.session, self.model, resource_id,
-                                  self.primary_key)
-        if primary_resource is None:
-            detail = 'No resource with ID {0}'.format(resource_id)
-            return error_response(404, detail=detail)
-        # Get the model of the specified relation.
-        related_model = get_related_model(self.model, relation_name)
-        if related_model is None:
-            detail = 'No such relation: {0}'.format(related_model)
-            return error_response(404, detail=detail)
-        # Determine if this is a to-one or a to-many relation.
-        if is_like_list(primary_resource, relation_name):
-            return self._get_collection_helper(resource=primary_resource,
-                                               relation_name=relation_name,
-                                               filters=filters, sort=sort,
-                                               group_by=group_by,
-                                               single=single)
-        else:
-            resource = getattr(primary_resource, relation_name)
-            return self._get_resource_helper(resource=resource,
-                                             primary_resource=primary_resource,
-                                             relation_name=relation_name)
-
-    def _get_resource(self, resource_id):
-        """Returns a response containing a single resource with the specified
-        ID.
-
-        For example, a request like::
-
-            GET /people/1
-
-        will fetch the person resource with ID 1.
-
-        In general, this method is called on requests of the form::
-
-            GET /<collection_name>/<resource_id>
-
-        """
-        for preprocessor in self.preprocessors['GET_RESOURCE']:
-            temp_result = preprocessor(resource_id=resource_id)
-            # Let the return value of the preprocessor be the new value of
-            # instid, thereby allowing the preprocessor to effectively specify
-            # which instance of the model to process on.
-            #
-            # We assume that if the preprocessor returns None, it really just
-            # didn't return anything, which means we shouldn't overwrite the
-            # instid.
-            if temp_result is not None:
-                resource_id = temp_result
-        # Get the resource with the specified ID.
-        resource = get_by(self.session, self.model, resource_id,
-                          self.primary_key)
-        if resource is None:
-            detail = 'No resource with ID {0}'.format(resource_id)
-            return error_response(404, detail=detail)
-        return self._get_resource_helper(resource)
-
-    def _get_collection(self):
-        """Returns a response containing a collection of resources of the type
-        specified by the ``model`` argument to the constructor of this class.
-
-        For example, a request like::
-
-            GET /people
-
-        will fetch a collection of people resources.
-
-        In general, this method is called on requests of the form::
-
-            GET /<collection_name>
-
-        Filtering, sorting, grouping, and pagination are applied to the
-        response in this method.
-
-        """
-        try:
-            filters, sort, group_by, single = self._collection_parameters()
-        except (TypeError, ValueError, OverflowError) as exception:
-            detail = 'Unable to decode filter objects as JSON list'
-            return error_response(400, cause=exception, detail=detail)
-        except SingleKeyError as exception:
-            detail = 'Invalid format for filter[single] query parameter'
-            return error_response(400, cause=exception, detail=detail)
-
-        for preprocessor in self.preprocessors['GET_COLLECTION']:
-            preprocessor(filters=filters, sort=sort, group_by=group_by,
-                         single=single)
-
-        return self._get_collection_helper(filters=filters, sort=sort,
-                                           group_by=group_by, single=single)
-
-    def get(self, resource_id, relation_name, related_resource_id):
-        """Returns the JSON document representing a resource or a collection of
-        resources.
-
-        If ``resource_id`` is ``None`` (that is, if the request is of the form
-        :http:get:`/people/`), this method returns a collection of resources.
-
-        Otherwise, if ``relation_name`` is ``None`` (that is, if the request is
-        of the form :http:get:`/people/1`), this method returns a resource with
-        the specified ID.
-
-        Otherwise, if ``related_resource_id`` is ``None`` (that is, if the
-        request is of the form :http:get:`/people/1/articles` or
-        :http:get:`/articles/1/author`), this method returns either a resource
-        in the case of a to-one relationship or a collection of resources in
-        the case of a to-many relationship.
-
-        Otherwise, if none of the arguments are ``None`` (that is, if the
-        request is of the form :http:get:`/people/1/articles/2`), this method
-        returns the particular resource in the to-many relationship with the
-        specified ID.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        if resource_id is None:
-            return self._get_collection()
-        if relation_name is None:
-            return self._get_resource(resource_id)
-        if related_resource_id is None:
-            return self._get_relation(resource_id, relation_name)
-        return self._get_related_resource(resource_id, relation_name,
-                                          related_resource_id)
-
-    def delete(self, resource_id):
-        """Deletes the resource with the specified ID.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        for preprocessor in self.preprocessors['DELETE_RESOURCE']:
-            temp_result = preprocessor(instance_id=resource_id)
-            # See the note under the preprocessor in the get() method.
-            if temp_result is not None:
-                resource_id = temp_result
-        was_deleted = False
-        instance = get_by(self.session, self.model, resource_id,
-                          self.primary_key)
-        if instance is None:
-            detail = 'No resource found with ID {0}'.format(resource_id)
-            return error_response(404, detail=detail)
-        self.session.delete(instance)
-        was_deleted = len(self.session.deleted) > 0
-        self.session.commit()
-        for postprocessor in self.postprocessors['DELETE_RESOURCE']:
-            postprocessor(was_deleted=was_deleted)
-        if not was_deleted:
-            detail = 'There was no instance to delete.'
-            return error_response(404, detail=detail)
-        return {}, 204
-
-    def post(self):
-        """Creates a new resource based on request data.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        # try to read the parameters for the model from the body of the request
-        try:
-            data = json.loads(request.get_data()) or {}
-        except (BadRequest, TypeError, ValueError, OverflowError) as exception:
-            detail = 'Unable to decode data'
-            return error_response(400, cause=exception, detail=detail)
-        # apply any preprocessors to the POST arguments
-        for preprocessor in self.preprocessors['POST']:
-            preprocessor(data=data)
-        if 'data' not in data:
-            detail = 'Resource must have a "data" key'
-            return error_response(400, detail=detail)
-        data = data['data']
-        # Convert the dictionary representation into an instance of the
-        # model.
-        if 'type' not in data:
-            detail = 'Must specify correct data type'
-            return error_response(400, detail=detail)
-        if 'id' in data and not self.allow_client_generated_ids:
-            detail = 'Server does not allow client-generated IDS'
-            return error_response(403, detail=detail)
-        type_ = data.pop('type')
-        if type_ != self.collection_name:
-            message = ('Type must be {0}, not'
-                       ' {1}').format(self.collection_name, type_)
-            return error_response(409, detail=message)
-        try:
-            instance = self.deserialize(data)
-            self.session.add(instance)
-            self.session.commit()
-        except DeserializationException as exception:
-            detail = 'Failed to deserialize object'
-            return error_response(400, cause=exception, detail=detail)
-        except self.validation_exceptions as exception:
-            return self._handle_validation_exception(exception)
-        fields = parse_sparse_fields()
-        fields_for_this = fields.get(self.collection_name)
-        # Get the dictionary representation of the new instance as it
-        # appears in the database.
-        try:
-            result = self.serialize(instance, only=fields_for_this)
-        except SerializationException as exception:
-            detail = 'Failed to serialize object'
-            return error_response(400, cause=exception, detail=detail)
-        # Determine the value of the primary key for this instance and
-        # encode URL-encode it (in case it is a Unicode string).
-        primary_key = primary_key_value(instance, as_string=True)
-        # The URL at which a client can access the newly created instance
-        # of the model.
-        url = '{0}/{1}'.format(request.base_url, primary_key)
-        # Provide that URL in the Location header in the response.
-        headers = dict(Location=url)
-        # Wrap the resulting object or list of objects under a 'data' key.
-        result = dict(data=result)
-        # Include any requested resources in a compound document.
-        to_include = self.resources_to_include(instance)
-        included = []
-        for included_resource in to_include:
-            type_ = collection_name(get_model(included_resource))
-            fields_for_this = fields.get(type_)
-            try:
-                serialized = self.serialize(included_resource,
-                                            only=fields_for_this)
-            except SerializationException as exception:
-                detail = 'Failed to serialize resource of type {0}'
-                detail = detail.format(type_)
-                return error_response(400, cause=exception, detail=detail)
-            included.append(serialized)
-        if included:
-            result['included'] = included
-        status = 201
-        for postprocessor in self.postprocessors['POST']:
-            postprocessor(result=result)
-        return result, status, headers
-
-    def _update_instance(self, instance, data):
-        """Updates the attributes and relationships of the specified instance
-        according to the elements in the `data` dictionary.
-
-        `instance` must be an instance of the SQLAlchemy model class specified
-        in the constructor of this class.
-
-        `data` must be a dictionary representation of a resource object as
-        described in the `Updating Resources`_ section of the JSON API
-        specification.
-
-        .. _Updating Resources: http://jsonapi.org/format/#crud-updating
-
-        """
-        # Update any relationships.
-        links = data.pop('relationships', {})
-        for linkname, link in links.items():
-            # The client is obligated by JSON API to provide linkage if
-            # the `links` attribute exists.
-            if 'data' not in link:
-                detail = 'relationship "{0}" is missing resource linkage'
-                detail = detail.format(linkname)
-                return error_response(400, detail=detail)
-            linkage = link['data']
-            related_model = get_related_model(self.model, linkname)
-            # If the client provided "null" for this relation, remove it by
-            # setting the attribute to ``None``.
-            if linkage is None:
-                setattr(instance, linkname, None)
-                continue
-            # If this is a to-many relationship, get all the related
-            # resources.
-            if isinstance(linkage, list):
-                # Replacement of a to-many relationship may have been disabled
-                # by the user.
-                if not self.allow_to_many_replacement:
-                    message = 'Not allowed to replace a to-many relationship'
-                    return error_response(403, detail=message)
-                # If this is left empty, the relationship will be zeroed.
-                newvalue = []
-                not_found = []
-                for rel in linkage:
-                    expected_type = collection_name(related_model)
-                    type_ = rel['type']
-                    if type_ != expected_type:
-                        detail = 'Type must be {0}, not {1}'
-                        detail = detail.format(expected_type, type_)
-                        return error_response(409, detail=detail)
-                    id_ = rel['id']
-                    inst = get_by(self.session, related_model, id_)
-                    if inst is None:
-                        not_found.append((id_, type_))
-                    else:
-                        newvalue.append(inst)
-                # If any of the requested to-many linkage objects do not exist,
-                # return an error response.
-                if not_found:
-                    detail = 'No object of type {0} found with ID {1}'
-                    errors = [error(detail=detail.format(t, i))
-                              for t, i in not_found]
-                    return errors_response(404, errors)
-            # Otherwise, it is a to-one relationship, so just get the single
-            # related resource.
-            else:
-                expected_type = collection_name(related_model)
-                type_ = linkage['type']
-                if type_ != expected_type:
-                    detail = 'Type must be {0}, not {1}'
-                    detail = detail.format(expected_type, type_)
-                    return error_response(409, detail=detail)
-                id_ = linkage['id']
-                inst = get_by(self.session, related_model, id_)
-                # If the to-one relationship resource does not exist, return an
-                # error response.
-                if inst is None:
-                    detail = 'No object of type {0} found with ID {1}'
-                    detail = detail.format(type_, id_)
-                    return error_response(404, detail=detail)
-                newvalue = inst
-            # Set the new value of the relationship.
-            try:
-                # TODO Here if there are any extra attributes in
-                # newvalue[inst], (1) get the secondary association object for
-                # that relation, then (2) set the extra attributes on that
-                # object.
-                setattr(instance, linkname, newvalue)
-            except self.validation_exceptions as exception:
-                return self._handle_validation_exception(exception)
-
-        # Now consider only the attributes to update.
-        data = data.pop('attributes', {})
-        # Check for any request parameter naming a column which does not exist
-        # on the current model.
-        for field in data:
-            if not has_field(self.model, field):
-                detail = "Model does not have field '{0}'".format(field)
-                return error_response(400, detail=detail)
-        # Special case: if there are any dates, convert the string form of the
-        # date into an instance of the Python ``datetime`` object.
-        data = strings_to_datetimes(self.model, data)
-        # Try to update all instances present in the query.
-        num_modified = 0
-        try:
-            if data:
-                for field, value in data.items():
-                    setattr(instance, field, value)
-                num_modified += 1
-            self.session.commit()
-        except self.validation_exceptions as exception:
-            return self._handle_validation_exception(exception)
-
-    def patch(self, resource_id):
-        """Updates the resource with the specified ID according to the request
-        data.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        # try to load the fields/values to update from the body of the request
-        try:
-            data = json.loads(request.get_data()) or {}
-        except (BadRequest, TypeError, ValueError, OverflowError) as exception:
-            # this also happens when request.data is empty
-            detail = 'Unable to decode data'
-            return error_response(400, cause=exception, detail=detail)
-        for preprocessor in self.preprocessors['PATCH_RESOURCE']:
-            temp_result = preprocessor(instance_id=resource_id, data=data)
-            # See the note under the preprocessor in the get() method.
-            if temp_result is not None:
-                resource_id = temp_result
-        # Get the instance on which to set the new attributes.
-        instance = get_by(self.session, self.model, resource_id,
-                          self.primary_key)
-        # If no instance of the model exists with the specified instance ID,
-        # return a 404 response.
-        if instance is None:
-            detail = 'No instance with ID {0} in model {1}'.format(resource_id,
-                                                                   self.model)
-            return error_response(404, detail=detail)
-        # Unwrap the data from the collection name key.
-        data = data.pop('data', {})
-        if 'type' not in data:
-            message = 'Must specify correct data type'
-            return error_response(400, detail=message)
-        if 'id' not in data:
-            message = 'Must specify resource ID'
-            return error_response(400, detail=message)
-        type_ = data.pop('type')
-        id_ = data.pop('id')
-        if type_ != self.collection_name:
-            message = ('Type must be {0}, not'
-                       ' {1}').format(self.collection_name, type_)
-            return error_response(409, detail=message)
-        if id_ != resource_id:
-            message = 'ID must be {0}, not {1}'.format(resource_id, id_)
-            return error_response(409, detail=message)
-        result = self._update_instance(instance, data)
-        # If result is not None, that means there was an error updating the
-        # resource.
-        if result is not None:
-            return result
-        # If we believe that the resource changes in ways other than the
-        # updates specified by the request, we must return 200 OK and a
-        # representation of the modified resource.
-        if self.changes_on_update:
-            result = dict(data=self.serialize(instance))
-            status = 200
-        else:
-            result = dict()
-            status = 204
-        # Perform any necessary postprocessing.
-        for postprocessor in self.postprocessors['PATCH_RESOURCE']:
-            postprocessor(result=result)
-        return result, status
-
-
-class RelationshipAPI(APIBase):
-    """Provides fetching, updating, and deleting from relationship URLs.
-
-    The endpoints provided by this class are of the form
-    ``/people/1/relationships/articles``, and the requests and responses
-    include **link objects**, as opposed to **resource objects**.
-
-    `session` and `model` are as described in the constructor of the
-    superclass. In addition to those described below, this constructor
-    also accepts all the keyword arguments of the constructor of the
-    superclass.
-
-    `allow_delete_from_to_many_relationships` is as described in
-    :meth:`APIManager.create_api`.
-
-    """
-
-    def __init__(self, session, model,
-                 allow_delete_from_to_many_relationships=False, *args, **kw):
-        super(RelationshipAPI, self).__init__(session, model, *args, **kw)
-        self.allow_delete_from_to_many_relationships = \
-            allow_delete_from_to_many_relationships
-
-    def collection_processor_type(self, *args, **kw):
-        return 'TO_MANY_RELATIONSHIP'
-
-    def resource_processor_type(self, *args, **kw):
-        return 'TO_ONE_RELATIONSHIP'
-
-    def use_resource_identifiers(self):
-        return True
-
-    def get(self, resource_id, relation_name):
-        """Fetches a to-one or to-many relationship from a resource.
-
-        If the specified relationship is a to-one relationship, this method
-        returns a link object. If it is a to-many relationship, it returns a
-        collection of link objects.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        for preprocessor in self.preprocessors['GET_RELATIONSHIP']:
-            temp_result = preprocessor(instance_id=resource_id,
-                                       relationship=relation_name)
-            # Let the return value of the preprocessor be the new value of
-            # instid, thereby allowing the preprocessor to effectively specify
-            # which instance of the model to process on.
-            #
-            # We assume that if the preprocessor returns None, it really just
-            # didn't return anything, which means we shouldn't overwrite the
-            # instid.
-            if temp_result is not None:
-                resource_id = temp_result
-        # get the instance of the "main" model whose ID is `resource_id`
-        primary_resource = get_by(self.session, self.model, resource_id,
-                                  self.primary_key)
-        if primary_resource is None:
-            detail = 'No resource with ID {0}'.format(resource_id)
-            return error_response(404, detail=detail)
-        if is_like_list(primary_resource, relation_name):
-            try:
-                filters, sort, group_by, single = self._collection_parameters()
-            except (TypeError, ValueError, OverflowError) as exception:
-                detail = 'Unable to decode filter objects as JSON list'
-                return error_response(400, cause=exception, detail=detail)
-            except SingleKeyError as exception:
-                detail = 'Invalid format for filter[single] query parameter'
-                return error_response(400, cause=exception, detail=detail)
-            return self._get_collection_helper(resource=primary_resource,
-                                               relation_name=relation_name,
-                                               filters=filters, sort=sort,
-                                               group_by=group_by,
-                                               single=single)
-        resource = getattr(primary_resource, relation_name)
-        return self._get_resource_helper(resource,
-                                         primary_resource=primary_resource,
-                                         relation_name=relation_name)
-
-    def post(self, resource_id, relation_name):
-        """Adds resources to a to-many relationship.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        # try to load the fields/values to update from the body of the request
-        try:
-            data = json.loads(request.get_data()) or {}
-        except (BadRequest, TypeError, ValueError, OverflowError) as exception:
-            # this also happens when request.data is empty
-            detail = 'Unable to decode data'
-            return error_response(400, cause=exception, detail=detail)
-        for preprocessor in self.preprocessors['POST_RELATIONSHIP']:
-            temp_result = preprocessor(instance_id=resource_id,
-                                       relation_name=relation_name, data=data)
-            # See the note under the preprocessor in the get() method.
-            if temp_result is not None:
-                resource_id, relation_name = temp_result
-        instance = get_by(self.session, self.model, resource_id,
-                          self.primary_key)
-        # If no instance of the model exists with the specified instance ID,
-        # return a 404 response.
-        if instance is None:
-            detail = 'No instance with ID {0} in model {1}'
-            detail = detail.format(resource_id, self.model)
-            return error_response(404, detail=detail)
-        # If no such relation exists, return a 404.
-        if not hasattr(instance, relation_name):
-            detail = 'Model {0} has no relation named {1}'
-            detail = detail.format(self.model, relation_name)
-            return error_response(404, detail=detail)
-        related_model = get_related_model(self.model, relation_name)
-        related_value = getattr(instance, relation_name)
-        # Unwrap the data from the request.
-        data = data.pop('data', {})
-        for rel in data:
-            if 'type' not in rel:
-                detail = 'Must specify correct data type'
-                return error_response(400, detail=detail)
-            if 'id' not in rel:
-                detail = 'Must specify resource ID'
-                return error_response(400, detail=detail)
-            type_ = rel['type']
-            # The type name must match the collection name of model of the
-            # relation.
-            if type_ != collection_name(related_model):
-                detail = ('Type must be {0}, not'
-                          ' {1}').format(collection_name(related_model), type_)
-                return error_response(409, detail=detail)
-            # Get the new objects to add to the relation.
-            new_value = get_by(self.session, related_model, rel['id'])
-            if new_value is None:
-                detail = ('No object of type {0} found with ID'
-                          ' {1}').format(type_, rel['id'])
-                return error_response(404, detail=detail)
-            # Don't append a new value if it already exists in the to-many
-            # relationship.
-            if new_value not in related_value:
-                try:
-                    related_value.append(new_value)
-                except self.validation_exceptions as exception:
-                    return self._handle_validation_exception(exception)
-        # TODO do we need to commit the session here?
-        #
-        #     self.session.commit()
-        #
-        # Perform any necessary postprocessing.
-        for postprocessor in self.postprocessors['POST_RELATIONSHIP']:
-            postprocessor()
-        return {}, 204
-
-    def patch(self, resource_id, relation_name):
-        """Updates to a to-one or to-many relationship.
-
-        If the relationship is a to-many relationship and this class was
-        instantiated with the ``allow_to_many_replacement`` keyword argument
-        set to ``False``, then this method returns a :http:status:`403`
-        response.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        # try to load the fields/values to update from the body of the request
-        try:
-            data = json.loads(request.get_data()) or {}
-        except (BadRequest, TypeError, ValueError, OverflowError) as exception:
-            # this also happens when request.data is empty
-            detail = 'Unable to decode data'
-            return error_response(400, cause=exception, detail=detail)
-        for preprocessor in self.preprocessors['PATCH_RELATIONSHIP']:
-            temp_result = preprocessor(instance_id=resource_id,
-                                       relation_name=relation_name, data=data)
-            # See the note under the preprocessor in the get() method.
-            if temp_result is not None:
-                resource_id, relation_name = temp_result
-        instance = get_by(self.session, self.model, resource_id,
-                          self.primary_key)
-        # If no instance of the model exists with the specified instance ID,
-        # return a 404 response.
-        if instance is None:
-            detail = 'No instance with ID {0} in model {1}'
-            detail = detail.format(resource_id, self.model)
-            return error_response(404, detail=detail)
-        # If no such relation exists, return a 404.
-        if not hasattr(instance, relation_name):
-            detail = 'Model {0} has no relation named {1}'
-            detail = detail.format(self.model, relation_name)
-            return error_response(404, detail=detail)
-        related_model = get_related_model(self.model, relation_name)
-        # related_value = getattr(instance, relation_name)
-
-        # Unwrap the data from the request.
-        data = data.pop('data', {})
-        # If the client sent a null value, we assume it wants to remove a
-        # to-one relationship.
-        if data is None:
-            if is_like_list(instance, relation_name):
-                detail = 'Cannot set null value on a to-many relationship'
-                return error_response(400, detail=detail)
-            setattr(instance, relation_name, None)
-        else:
-            # If this is a list, we assume the client is trying to set a
-            # to-many relationship.
-            if isinstance(data, list):
-                # Replacement of a to-many relationship may have been disabled
-                # on the server-side by the user.
-                if not self.allow_to_many_replacement:
-                    detail = 'Not allowed to replace a to-many relationship'
-                    return error_response(403, detail=detail)
-                replacement = []
-                for rel in data:
-                    if 'type' not in rel:
-                        detail = 'Must specify correct data type'
-                        return error_response(400, detail=detail)
-                    if 'id' not in rel:
-                        detail = 'Must specify resource ID or IDs'
-                        return error_response(400, detail=detail)
-                    type_ = rel['type']
-                    # The type name must match the collection name of model of
-                    # the relation.
-                    if type_ != collection_name(related_model):
-                        detail = 'Type must be {0}, not {1}'
-                        detail = detail.format(collection_name(related_model),
-                                               type_)
-                        return error_response(409, detail=detail)
-                    id_ = rel['id']
-                    obj = get_by(self.session, related_model, id_)
-                    replacement.append(obj)
-            # Otherwise, we assume the client is trying to set a to-one
-            # relationship.
-            else:
-                if 'type' not in data:
-                    detail = 'Must specify correct data type'
-                    return error_response(400, detail=detail)
-                if 'id' not in data:
-                    detail = 'Must specify resource ID or IDs'
-                    return error_response(400, detail=detail)
-                type_ = data['type']
-                # The type name must match the collection name of model of the
-                # relation.
-                if type_ != collection_name(related_model):
-                    detail = ('Type must be {0}, not'
-                              ' {1}').format(collection_name(related_model),
-                                             type_)
-                    return error_response(409, detail=detail)
-                id_ = data['id']
-                replacement = get_by(self.session, related_model, id_)
-            # If the to-one relationship resource or any of the to-many
-            # relationship resources do not exist, return an error response.
-            if replacement is None:
-                detail = ('No object of type {0} found'
-                          ' with ID {1}').format(type_, id_)
-                return error_response(404, detail=detail)
-            if (isinstance(replacement, list)
-                and any(value is None for value in replacement)):
-                not_found = (rel for rel, value in zip(data, replacement)
-                             if value is None)
-                detail = 'No object of type {0} found with ID {1}'
-                errors = [error(detail=detail.format(rel['type'], rel['id']))
-                          for rel in not_found]
-                return errors_response(404, errors)
-            # Finally, set the relationship to have the new value.
-            try:
-                setattr(instance, relation_name, replacement)
-            except self.validation_exceptions as exception:
-                return self._handle_validation_exception(exception)
-        # TODO do we need to commit the session here?
-        #
-        #     self.session.commit()
-        #
-        # Perform any necessary postprocessing.
-        for postprocessor in self.postprocessors['PATCH_RELATIONSHIP']:
-            postprocessor()
-        return {}, 204
-
-    def delete(self, resource_id, relation_name):
-        """Deletes resources from a to-many relationship.
-
-        If this class was instantiated with the
-        ``allow_delete_from_to_many_relationships`` keyword argument set to
-        ``False``, then this method returns a :http:status:`403` response.
-
-        The request documents, response documents, and status codes are in the
-        format specified by the JSON API specification.
-
-        """
-        if not self.allow_delete_from_to_many_relationships:
-            detail = 'Not allowed to delete from a to-many relationship'
-            return error_response(403, detail=detail)
-        # try to load the fields/values to update from the body of the request
-        try:
-            data = json.loads(request.get_data()) or {}
-        except (BadRequest, TypeError, ValueError, OverflowError) as exception:
-            # this also happens when request.data is empty
-            detail = 'Unable to decode data'
-            return error_response(400, cause=exception, detail=detail)
-        was_deleted = False
-        for preprocessor in self.preprocessors['DELETE_RELATIONSHIP']:
-            temp_result = preprocessor(instance_id=resource_id,
-                                       relation_name=relation_name)
-            # See the note under the preprocessor in the get() method.
-            if temp_result is not None:
-                resource_id = temp_result
-        instance = get_by(self.session, self.model, resource_id,
-                          self.primary_key)
-        # If no such relation exists, return an error to the client.
-        if not hasattr(instance, relation_name):
-            detail = 'No such link: {0}'.format(relation_name)
-            return error_response(404, detail=detail)
-        # We assume that the relation is a to-many relation.
-        related_model = get_related_model(self.model, relation_name)
-        related_type = collection_name(related_model)
-        relation = getattr(instance, relation_name)
-        data = data.pop('data')
-        not_found = []
-        to_remove = []
-        for rel in data:
-            if 'type' not in rel:
-                detail = 'Must specify correct data type'
-                return error_response(400, detail=detail)
-            if 'id' not in rel:
-                detail = 'Must specify resource ID'
-                return error_response(400, detail=detail)
-            type_ = rel['type']
-            id_ = rel['id']
-            if type_ != related_type:
-                detail = ('Conflicting type: expected {0} but got type {1} for'
-                          ' linkage object with ID {2}')
-                detail = detail.format(related_type, type_, id_)
-                return error_response(409, detail=detail)
-            resource = get_by(self.session, related_model, id_)
-            if resource is None:
-                not_found.append((type_, id_))
-            else:
-                to_remove.append(resource)
-        if not_found:
-            detail = 'No resource of type {0} and ID {1} found'
-            errors = [error(detail=detail.format(t, i)) for t, i in not_found]
-            return errors_response(404, errors)
-        # Remove each of the resources from the relation (if they are not
-        # already absent).
-        for resource in to_remove:
-            try:
-                relation.remove(resource)
-            except ValueError:
-                # The JSON API specification requires that we silently
-                # ignore requests to delete resources that are already
-                # missing from a to-many relation.
-                pass
-        was_deleted = len(self.session.dirty) > 0
-        self.session.commit()
-        for postprocessor in self.postprocessors['DELETE_RELATIONSHIP']:
-            postprocessor(was_deleted=was_deleted)
-        if not was_deleted:
-            detail = 'There was no instance to delete'
-            return error_response(404, detail=detail)
-        return {}, 204
